Update submission certificate to match shim compilation cert
Mistakenly, the certificate used in review submission was not the same
as the one used to compile the shim. This should not affect the
submission as the compiled binary is still the same, but the certificate
present in the review submission is now updated to be aligned with the
shim build process.